dispatch parsed token on log in
