+++ src/store/store.js
@@ -3,13 +3,12 @@ import thunk from "redux-thunk";
 import { combineReducers } from 'redux';
 import { counterReducer } from './reducers/counterReducers';
 import { loginReducer } from './reducers/loginReducers';
-import { languageReducer, roleReducer } from './reducers/headerReducers';
+import { languageReducer } from './reducers/headerReducers';
 
 const rootReducers = combineReducers({
   counter: counterReducer,
   login: loginReducer,
-  language: languageReducer,
-  role: roleReducer,
+  language: languageReducer
 });
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 const store = createStore(rootReducers, composeEnhancers(applyMiddleware(thunk)));
